+++ src/tile.cpp
@@ -1094,8 +1094,8 @@ void Tile::removeThing(Thing* thing, uint32_t count)
 		}
 
 		if (itemType.stackable && count != item->getItemCount()) {
-			uint8_t newCount =
-			    static_cast<uint8_t>(std::max<int32_t>(0, static_cast<int32_t>(item->getItemCount() - count)));
+			uint16_t newCount =
+			    static_cast<uint16_t>(std::max<int32_t>(0, static_cast<int32_t>(item->getItemCount() - count)));
 			item->setItemCount(newCount);
 			onUpdateTileItem(item, itemType, item, itemType);
 		} else {
